requirements.txt
amqp==2.6.0
billiard==3.6.3.0
blinker==1.4
celery==4.4.4
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
dnspython==1.16.0
email-validator==1.1.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
future==0.18.2
idna==2.9
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.10
MarkupSafe==1.1.1
numpy==1.18.5
oauthlib==3.1.0
pandas==1.0.4
pendulum==2.1.0
psycopg2==2.8.5
pycparser==2.20
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
pytzdata==2019.3
redis==3.5.3
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
-e git+https://github.com/limodou/uliweb3@34b3ab5ead23950df18ebd2d6e10d105e8296640#egg=Uliweb3
urllib3==1.25.9
vine==1.3.0
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
